versionNumber
lavet double om til int

+++ Library/FileVersion.cs
@@ -9,11 +9,11 @@ namespace Library
     {
         public int Id { get; set; }
         public string FilePath { get; set; }
-        public double VersionNumber { get; set; }
+        public int VersionNumber { get; set; }
         public DateTime TimeStamp { get; set; }
         public User Owner { get; set; }
 
-        public FileVersion(string filePath, double versionNumber, User owner)
+        public FileVersion(string filePath, int versionNumber, User owner)
         {
             FilePath = filePath;
             VersionNumber = versionNumber;
@@ -21,7 +21,7 @@ namespace Library
             Owner = owner;
         }
 
-        public FileVersion(int id, string filePath, double versionNumber, User owner)
+        public FileVersion(int id, string filePath, int versionNumber, User owner)
         {
             Id = id;
             FilePath = filePath;
@@ -30,7 +30,7 @@ namespace Library
             Owner = owner;
         }
 
-        public FileVersion(int id, string filePath, double versionNumber, DateTime timeStamp, User owner)
+        public FileVersion(int id, string filePath, int versionNumber, DateTime timeStamp, User owner)
         {
             Id = id;
             FilePath = filePath;
